add noZFS option to fio.sh, improved fioparse.sh output
fioparse.sh  no prompts for test name when outputting R format
fiog.r also creates PNG files with and without histograms overlaid

+++ fio.sh
@@ -3,9 +3,11 @@
 #
 #  DIRECT I/O :
 #
+#  direct I/O is used except on ZFS
+#  DelphixOS uses ZFS 
+#  thus
 #  direct is set to 0 if it's detected that fio.sh is running on Delphix
-#  environment variable FDIO=1
-#  will direct=1 ven when trunning on Delphix O/S
+#  environment variable FDIO=1 will force direct=1 when running on Delphix O/S
 #  direct=1 doesn't seem to work on opensolaris even
 #  when opensolaris is the NFS client
 #
@@ -76,13 +78,14 @@ OPTIONS:
    -u #users       force test to only use this many users
    -l blocksize    force test to only use this blocksize in KB, ie 1-1024 
    -e recordsize   use this recordsize if/when creating the zfs file system, default 8K
+   -z              turn off default ZFS setup, ie creating a domain etc
 
        example
                   fio.sh ./fio.opensolaris /domain0/fiotest  -t rand_read -s 10 -m 1000 -f
 EOF
 }
 
-while getopts hb:r:e:d:o:it:s:l:u:m:f OPTION
+while getopts hb:r:ce:d:o:it:s:l:u:m:fz OPTION
 do
      case $OPTION in
          h)
@@ -132,6 +135,9 @@ do
          t)
              TESTS=$OPTARG
              ;;
+         z)
+             NOZFS=1
+             ;;
          ?)
              usage
              exit
@@ -139,7 +145,10 @@ do
      esac
 done
 
-
+#
+#  BEGIN {
+#  DTRACE SCRIPT GENERATION FUNCTIONS
+#
 
 dtrace_begin()
 {
@@ -263,11 +272,14 @@ offsets()
      done
 }
 
-# if there is no filename specified
-# then fio will use a file per processes 
-# instead of a single file
-# the filenames will be generated
-# each generated file will get the same size 
+#
+#  DTRACE SCRIPT GENERATION FUNCTIONS
+#  END }
+#
+
+#
+#  CUSTOM  WORKFILE SETUP  
+#
 if [ x$FILENAME == x ] ; then
     SIZE="size=100m"
     if [ x$MB == x1 ]; then
@@ -277,12 +289,18 @@ if [ x$FILENAME == x ] ; then
 fi
 
 
+#
+#  OUTPUT DIRECTORY SETUP
+#
 mkdir $OUTPUT > /dev/null 2>&1
 if [ ! -d $OUTPUT ]; then 
   echo "directory $OUTPUT does not exist"
   exit
 fi
 
+#
+#   DELPHIX OS DETECTION
+#
 if [ -f /etc/delphix/version ]  ; then 
    DIRECT=0
    # if running on Delphix, then collect DTrace I/O info
@@ -290,6 +308,9 @@ if [ -f /etc/delphix/version ]  ; then
    DTRACE2=" ' -s fio.d  "
 fi
 
+#
+#  DEFINE TESTS TO RUN
+#
 all="randrw read write readrand"
 all="readrand write read "
 if [ $TESTS = "all" ] ; then
@@ -299,7 +320,6 @@ else
 fi
 
 DIRECT=${FDIO:-$DIRECT}
-
 echo "configuration: "
 echo "    binary=$BINARY"
 echo "    work directory=$DIRECTORY"
@@ -314,13 +334,22 @@ echo "    recordsize =$RECORDSIZE"
 echo "    filename (blank if multiple files)=\"$FILENAME\""
 echo "    size per file of multiple files=\"$SIZE\""
 
-if [ -f /etc/delphix/version ] && [ $RAW -eq 0 ] ; then 
+
+#
+#  ZFS SETUP
+#
+#   if running on DELPHIX, set up ZFS files system
+#   (this options could be made a command line flag
+#    for other systems that might want to test ZFS)
+#
+#  if running on DELPHIX and not using RAW devices
+#  and NOZFS has not been set
+if [ -f /etc/delphix/version ] && [ $RAW -eq 0 ] && [ $NOZFS -eq 0 ]; then 
    if [  -f fio.d ]; then 
      if [ ! -f fio.d ]; then
        mv fio.d fio.d.orig
      fi
    fi
-
    # DIRECTORY=/domain0/fiotest
    FILESYSTEM=`echo $DIRECTORY | sed -e 's;^/;;' `
    DOMAIN=`echo $FILESYSTEM    | sed -e 's;/.*;;' `
@@ -393,6 +422,10 @@ if [ -f /etc/delphix/version ] && [ $RAW -eq 0 ] ; then
    cat $OUTPUT/setup.txt | grep -v PROPERTY | sed -e 's/^/   /' 
 fi 
 
+
+#
+#  DIRECTORY to create I/O test file
+#
 if [ ! -d $DIRECTORY ]; then 
   echo "directory $DIRECTORY does not exist"
   exit
@@ -406,6 +439,10 @@ if [ $FORCE = "n" ] ; then
   fi
 fi
 
+#
+#  DTRACE SETUP 
+#   (should be added as a command line option)
+#
 if [ -f /etc/delphix/version ]  ; then 
    dtrace_begin > fio.d
    if [ $RAW == 1 ] ; then 
@@ -425,12 +462,14 @@ if [ -f /etc/delphix/version ]  ; then
 fi
 
 
-
+#
+#  CREATE FILE for I/O tests 
+#
  if [ ! -f $DIRECTORY/$FILE ]  ||  [ $CREATE == 1  ]; then
    # tar cvf - /opt/delphix/server > /domain0/fiotest/fiodata
-   if [ $RAW == 0 ] ; then
+   if [ $RAW == 0 ] ; then  # BEGIN NOT RAW DEVICE {
     echo "CREATE=$DIRECTORY/$FILE"
-    if [ -f $SEEDFILE ] ; then 
+    if [ -f $SEEDFILE ] ; then # BEGIN SEEDFILE {
        echo "seed file found, using $SEEDFILE"
        loops=0
        while [[ $loops -le $MEGABYTES ]] ; do
@@ -466,11 +505,12 @@ fi
        done
        rm dd.out.$$
        rm /tmp/fio.$$
-   fi
+   fi # } END SEEDFILE
    echo 
    echo "file creation finished"
-  fi
+  fi # } END NOT RAW DEVICE 
  fi
+
 
 if [ $RAW -eq 0 ]; then
    cmd="ls -l $DIRECTORY/$FILE "
